requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
google-generativeai
google-auth-oauthlib
google-api-python-client
requests